// src/types/business.ts
export type Business = {
  id: number
  place_id: string | null
  name: string | null
  address: string | null
  phone: string | null
  website: string | null
  rating: number | null
  total_ratings: number | null
  category: string | null
  city: string | null
  types: string | null
  status: string | null
  website_type: string | null
  web_scrape_status: string | null
  email: string | null
  ig_handle: string | null
  contacted: number | null
  created_at: number | null
  updated_at: number | null
}